Uploaded excel document
This is the basic outline for the waste tracking application I plan to develop. It will work fine as an excel document, however I would like to develop this into an application that can have cross facility functionality and is relatively easy for the end user to operate without breaking.